requirements.txt
mcp[cli]>=1.0.0
playwright>=1.40.0
Pillow>=10.0.0
numpy>=1.24.0
